+++ examples/sqlite_remember_bot/src/transitions.rs
@@ -1,5 +1,5 @@
 use teloxide::prelude::*;
-use teloxide_macros::teloxide;
+use teloxide::macros::teloxide;
 
 use super::states::*;